+++ hi-ml-histopathology/testhisto/testhisto/models/test_deepmil.py
@@ -4,7 +4,7 @@
 #  ------------------------------------------------------------------------------------------
 
 import os
-from typing import Any, Callable, Dict, Iterable, List, Type
+from typing import Any, Callable, Dict, Iterable, List, Optional, Type, Tuple
 from unittest.mock import MagicMock
 
 import pytest
@@ -14,10 +14,8 @@ from torch.utils.data._utils.collate import default_collate
 from torchvision.models import resnet18
 
 from health_ml.lightning_container import LightningContainer
-from health_ml.networks.layers.attention_layers import (
-    AttentionLayer,
-    GatedAttentionLayer,
-)
+from health_ml.networks.layers.attention_layers import AttentionLayer
+
 
 from histopathology.configs.classification.DeepSMILECrck import DeepSMILECrck
 from histopathology.configs.classification.DeepSMILEPanda import DeepSMILEPanda
@@ -33,32 +31,41 @@ def get_supervised_imagenet_encoder() -> TileEncoder:
     return ImageNetEncoder(feature_extraction_model=resnet18, tile_size=224)
 
 
-@pytest.mark.parametrize("n_classes", [1, 3])
-@pytest.mark.parametrize("pooling_layer", [AttentionLayer, GatedAttentionLayer])
-@pytest.mark.parametrize("batch_size", [1, 2])
-@pytest.mark.parametrize("max_bag_size", [1, 3])
-@pytest.mark.parametrize("pool_hidden_dim", [1, 4])
-@pytest.mark.parametrize("pool_out_dim", [1, 5])
-def test_lightningmodule(
+def get_attention_pooling_layer(num_encoding: int = 512,
+                                pool_out_dim: int = 1) -> Tuple[Type[nn.Module], int]:
+
+    pool_hidden_dim = 5  # different dimensions get tested in test_attentionlayers.py
+    pooling_layer = AttentionLayer(num_encoding,
+                                   pool_hidden_dim,
+                                   pool_out_dim)
+
+    num_features = num_encoding * pool_out_dim
+    return pooling_layer, num_features
+
+
+def _test_lightningmodule(
     n_classes: int,
-    pooling_layer: Callable[[int, int, int], nn.Module],
     batch_size: int,
     max_bag_size: int,
-    pool_hidden_dim: int,
     pool_out_dim: int,
+    dropout_rate: Optional[float],
 ) -> None:
 
     assert n_classes > 0
 
     # hard-coded here to avoid test explosion; correctness of other encoders is tested elsewhere
     encoder = get_supervised_imagenet_encoder()
+
+    # hard-coded here to avoid test explosion; correctness of other pooling layers is tested elsewhere
+    pooling_layer, num_features = get_attention_pooling_layer(pool_out_dim=pool_out_dim)
+
     module = DeepMILModule(
         encoder=encoder,
         label_column="label",
         n_classes=n_classes,
         pooling_layer=pooling_layer,
-        pool_hidden_dim=pool_hidden_dim,
-        pool_out_dim=pool_out_dim,
+        num_features=num_features,
+        dropout_rate=dropout_rate,
     )
 
     bag_images = rand([batch_size, max_bag_size, *module.encoder.input_dim])
@@ -73,7 +80,7 @@ def test_lightningmodule(
         bag_labels_list.append(module.get_bag_label(labels))
         logit, attn = module(bag)
         assert logit.shape == (1, n_classes)
-        assert attn.shape == (module.pool_out_dim, max_bag_size)
+        assert attn.shape == (pool_out_dim, max_bag_size)
         bag_logits_list.append(logit.view(-1))
         bag_attn_list.append(attn)
 
@@ -93,31 +100,51 @@ def test_lightningmodule(
 
     probs = module.activation_fn(bag_logits)
     assert ((probs >= 0) & (probs <= 1)).all()
-    if n_classes > 1:
-        assert probs.shape == (batch_size, n_classes)
-    else:
-        assert probs.shape[0] == batch_size
 
     if n_classes > 1:
-        preds = argmax(probs, dim=1)
+        predlabels = argmax(probs, dim=1)
     else:
-        preds = round(probs)
-    assert preds.shape[0] == batch_size
+        predlabels = round(probs)
 
+    predlabels = predlabels.view(-1, 1)
+    assert predlabels.shape[0] == batch_size
+    assert probs.shape == (batch_size, n_classes)
+
+    bag_labels = bag_labels.view(-1, 1)
+    if n_classes == 1:
+        probs = probs.squeeze(dim=1)
     for metric_name, metric_object in module.train_metrics.items():
-        if metric_name == MetricsKey.CONF_MATRIX or metric_name == MetricsKey.AUROC:
+        if metric_name == MetricsKey.CONF_MATRIX:
             continue
-        if batch_size > 1:
-            score = metric_object(preds.view(-1, 1), bag_labels.view(-1, 1))
-            assert torch.all(score >= 0)
-            assert torch.all(score <= 1)
+        score = metric_object(probs, bag_labels.view(batch_size,))
+        assert torch.all(score >= 0)
+        assert torch.all(score <= 1)
+
+
+@pytest.mark.parametrize("n_classes", [1, 3])
+@pytest.mark.parametrize("batch_size", [1, 15])
+@pytest.mark.parametrize("max_bag_size", [1, 7])
+@pytest.mark.parametrize("pool_out_dim", [1, 6])
+@pytest.mark.parametrize("dropout_rate", [None, 0.5])
+def test_lightningmodule_attention(
+    n_classes: int,
+    batch_size: int,
+    max_bag_size: int,
+    pool_out_dim: int,
+    dropout_rate: Optional[float],
+) -> None:
+    _test_lightningmodule(n_classes=n_classes,
+                          batch_size=batch_size,
+                          max_bag_size=max_bag_size,
+                          pool_out_dim=pool_out_dim,
+                          dropout_rate=dropout_rate)
 
 
 def validate_metric_inputs(scores: torch.Tensor, labels: torch.Tensor) -> None:
     def is_integral(x: torch.Tensor) -> bool:
         return (x == x.long()).all()  # type: ignore
 
-    assert scores.shape == labels.shape
+    assert labels.shape == (scores.shape[0], )
     assert torch.is_floating_point(scores), "Received scores with integer dtype"
     assert not is_integral(scores), "Received scores with integral values"
     assert is_integral(labels), "Received labels with floating-point values"
@@ -130,13 +157,20 @@ def add_callback(fn: Callable, callback: Callable) -> Callable:
     return wrapper
 
 
-def test_metrics() -> None:
+@pytest.mark.parametrize("n_classes", [1, 3])
+def test_metrics(n_classes: int) -> None:
     input_dim = (128,)
+
+    # hard-coded here to avoid test explosion; correctness of other pooling layers is tested elsewhere
+    pooling_layer, num_features = get_attention_pooling_layer(num_encoding=input_dim[0],
+                                                              pool_out_dim=1)
+
     module = DeepMILModule(
         encoder=IdentityEncoder(input_dim=input_dim),
         label_column=TilesDataset.LABEL_COLUMN,
-        n_classes=1,
-        pooling_layer=AttentionLayer,
+        n_classes=n_classes,
+        pooling_layer=pooling_layer,
+        num_features=num_features
     )
 
     # Patching to enable running the module without a Trainer object
@@ -145,7 +179,10 @@ def test_metrics() -> None:
 
     batch_size = 20
     bag_size = 5
-    class_weights = torch.tensor([.8, .2])
+    if n_classes > 1:
+        class_weights = torch.rand(n_classes)
+    else:
+        class_weights = torch.tensor([0.8, 0.2])
     bags: List[Dict] = []
     for slide_idx in range(batch_size):
         bag_label = torch.multinomial(class_weights, 1)
@@ -177,40 +214,8 @@ def test_metrics() -> None:
 
     for key, metric_obj in module_metrics_dict.items():
         value = metric_obj.compute()
-        expected_value = independent_metrics_dict[key](predicted_probs, true_labels)
+        expected_value = independent_metrics_dict[key](predicted_probs, true_labels.view(batch_size,))
         assert torch.allclose(value, expected_value), f"Discrepancy in '{key}' metric"
-
-    # ================
-    # Test that thresholded metrics (e.g. accuracy, precision, etc.) change as the threshold is varied.
-    # If they don't, it suggests the inputs are hard labels instead of continuous scores.
-    thresholded_metrics_keys = [key for key, metric in module_metrics_dict.items()
-                                if hasattr(metric, 'threshold')]
-
-    def set_metrics_threshold(metrics_dict: Any, threshold: float) -> None:
-        for key in thresholded_metrics_keys:
-            metrics_dict[key].threshold = threshold
-
-    def reset_metrics(metrics_dict: Any) -> None:
-        for metric_obj in metrics_dict.values():
-            metric_obj.reset()
-
-    low_threshold, high_threshold = torch.quantile(predicted_probs, torch.tensor([0.1, 0.9]))
-
-    reset_metrics(module_metrics_dict)
-    set_metrics_threshold(module_metrics_dict, threshold=low_threshold)
-    _ = module.test_step(batch, 0)
-    results_low_threshold = {key: module_metrics_dict[key].compute()
-                             for key in thresholded_metrics_keys}
-
-    reset_metrics(module_metrics_dict)
-    set_metrics_threshold(module_metrics_dict, threshold=high_threshold)
-    _ = module.test_step(batch, 0)
-    results_high_threshold = {key: module_metrics_dict[key].compute()
-                              for key in thresholded_metrics_keys}
-
-    for key in thresholded_metrics_keys:
-        assert not torch.allclose(results_low_threshold[key], results_high_threshold[key]), \
-            f"Got same value for '{key}' metric with low and high thresholds"
 
 
 def move_batch_to_expected_device(batch: Dict[str, List], use_gpu: bool) -> Dict:
@@ -286,13 +291,16 @@ def test_container(container_type: Type[LightningContainer], use_gpu: bool) -> N
 def test_class_weights_binary() -> None:
     class_weights = Tensor([0.5, 3.5])
     n_classes = 1
+
+    # hard-coded here to avoid test explosion; correctness of other pooling layers is tested elsewhere
+    pooling_layer, num_features = get_attention_pooling_layer(pool_out_dim=1)
+
     module = DeepMILModule(
         encoder=get_supervised_imagenet_encoder(),
         label_column="label",
         n_classes=n_classes,
-        pooling_layer=AttentionLayer,
-        pool_hidden_dim=5,
-        pool_out_dim=1,
+        pooling_layer=pooling_layer,
+        num_features=num_features,
         class_weights=class_weights,
     )
     logits = Tensor(randn(1, n_classes))
@@ -311,13 +319,16 @@ def test_class_weights_binary() -> None:
 def test_class_weights_multiclass() -> None:
     class_weights = Tensor([0.33, 0.33, 0.33])
     n_classes = 3
+
+    # hard-coded here to avoid test explosion; correctness of other pooling layers is tested elsewhere
+    pooling_layer, num_features = get_attention_pooling_layer(pool_out_dim=1)
+
     module = DeepMILModule(
         encoder=get_supervised_imagenet_encoder(),
         label_column="label",
         n_classes=n_classes,
-        pooling_layer=AttentionLayer,
-        pool_hidden_dim=5,
-        pool_out_dim=1,
+        pooling_layer=pooling_layer,
+        num_features=num_features,
         class_weights=class_weights,
     )
     logits = Tensor(randn(1, n_classes))